% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_all_values_positive}
\alias{are_all_values_positive}
\title{\code{are_all_values_positive}: check for strictly positive values in object}
\usage{
are_all_values_positive(x)
}
\arguments{
\item{x}{A numeric object}
}
\value{
A boolean. TRUE if all values are positive, FALSE otherwise
}
\description{
\code{are_all_values_positive} test presumably a vector, but any numeric object, for strictly positive values
}
