% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{authors}
\alias{authors}
\title{Authors}
\format{
A matrix that contains the frequency of punctuation usage of 6 authors x 3 punctuations.
}
\usage{
authors
}
\description{
Data used for illustrative purposes in Abdi, H. & Williams, L.J. (2010). Correspondence analysis. In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.): Encyclopedia of Research Design. Thousand Oaks (CA): Sage. pp. 267-278.
}
\keyword{datasets}
