% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{beer.tasting.notes}
\alias{beer.tasting.notes}
\title{Personal beer tasting notes}
\format{A list called 'beer.tasting.notes' with four matrices:
\describe{
 \item{data}{Data matrix. Tasting notes (ratings) of 38 different beers (rows) described by 16 different flavor profiles (columns).}
 \item{brewery.design}{Design matrix. Source brewery of 38 different beers (rows) across 26 breweries (columns).}
 \item{style.design}{Design matrix. Style of 38 different beers (rows) across 20 styles (columns) (styles as listed from Beer Advocate website).}
 \item{sup.data}{Supplementary data matrix. ABV and overall preference ratings of 38 beers described by two features (ABV & overall) in original value and rounded value.}
}}
\source{
Jenny Rieck and Derek Beaton laboriously ``collected'' these data for ``experimental purposes''.
}
\usage{
beer.tasting.notes
}
\description{
Tasting notes, preferences, breweries and styles of 38 different craft beers from various breweries, across various styles.
}
\references{
http://www.beeradvocate.com
}
\author{
Jenny Rieck and Derek Beaton
}
\keyword{datasets}
