% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplssvd.R
\name{gplssvd}
\alias{gplssvd}
\title{Generalized partial least squares singular value decomposition}
\usage{
gplssvd(X, Y, XLW, YLW, XRW, YRW, k = 0, tol = .Machine$double.eps)
}
\arguments{
\item{X}{a data matrix}

\item{Y}{a data matrix}

\item{XLW}{\bold{X}'s \bold{L}eft \bold{W}eights -- the constraints applied to the left side (rows) of the \code{X} matrix and thus left singular vectors.}

\item{YLW}{\bold{Y}'s \bold{L}eft \bold{W}eights -- the constraints applied to the left side (rows) of the \code{Y} matrix and thus left singular vectors.}

\item{XRW}{\bold{X}'s \bold{R}ight \bold{W}eights -- the constraints applied to the right side (columns) of the \code{X} matrix and thus right singular vectors.}

\item{YRW}{\bold{Y}'s \bold{R}ight \bold{W}eights -- the constraints applied to the right side (columns) of the \code{Y} matrix and thus right singular vectors.}

\item{k}{total number of components to return though the full variance will still be returned (see \code{d_full}). If 0, the full set of components are returned.}

\item{tol}{default is .Machine$double.eps. A tolerance level for eliminating effectively zero (small variance), negative, imaginary eigen/singular value components.}
}
\value{
A list with thirteen elements:
\item{d_full}{A vector containing the singular values of DAT above the tolerance threshold (based on eigenvalues).}
\item{l_full}{A vector containing the eigen values of DAT above the tolerance threshold (\code{tol}).}
\item{d}{A vector of length \code{min(length(d_full), k)} containing the retained singular values}
\item{l}{A vector of length \code{min(length(l_full), k)} containing the retained eigen values}
\item{u}{Left (rows) singular vectors. Dimensions are \code{nrow(DAT)} by k.}
\item{p}{Left (rows) generalized singular vectors.}
\item{fi}{Left (rows) component scores.}
\item{lx}{Latent variable scores for rows of \code{X}}
\item{v}{Right (columns) singular vectors.}
\item{q}{Right (columns) generalized singular vectors.}
\item{fj}{Right (columns) component scores.}
\item{ly}{Latent variable scores for rows of \code{Y}}
}
\description{
\code{gplssvd} takes in left (\code{XLW}, \code{YLW}) and right (\code{XRW}, \code{YRW}) constraints (usually diagonal matrices, but any positive semi-definite matrix is fine) that are applied to each data matrix (\code{X} and \code{Y}), respectively.
  Right constraints for each data matrix are used for the orthogonality conditions.
}
\examples{

 # Three "two-table" technique examples
 data(wine)
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)

 ## Partial least squares (correlation)
 pls.res <- gplssvd(X, Y)


 ## Canonical correlation analysis (CCA)
 ### NOTE:
 #### This is not "traditional" CCA because of the generalized inverse.
 #### However the results are the same as standard CCA when data are not rank deficient.
 #### and this particular version uses tricks to minimize memory & computation
 cca.res <- gplssvd(
     X = MASS::ginv(t(X)),
     Y = MASS::ginv(t(Y)),
     XRW=crossprod(X),
     YRW=crossprod(Y)
 )


 ## Reduced rank regression (RRR) a.k.a. redundancy analysis (RDA)
 ### NOTE:
 #### This is not "traditional" RRR because of the generalized inverse.
 #### However the results are the same as standard RRR when data are not rank deficient.
 rrr.res <- gplssvd(
     X = MASS::ginv(t(X)),
     Y = Y,
     XRW=crossprod(X)
 )

}
\seealso{
\code{\link{tolerance_svd}}, \code{\link{geigen}} and \code{\link{gsvd}}
}
\author{
Derek Beaton
}
\keyword{multivariate}
