% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.exponent.R
\name{\%^\%}
\alias{\%^\%}
\title{Matrix exponentiation}
\usage{
x \%^\% power
}
\arguments{
\item{x}{data matrix}

\item{power}{the power to raise \code{x} by (e.g., 2 is squared)}
}
\value{
\code{x} raised to an arbitrary \code{power}
}
\description{
takes in a matrix and will compute raise that matrix to some arbitrary power via the singular value decomposition.
 Additionally, the matrix can be computed for a lower rank estimate of the matrix.
}
\examples{
 data(wine)
 X <- as.matrix(wine$objective)
 X \%^\% 2 # power of 2
 X \%^\% -1 # (generalized) inverse

}
\seealso{
\code{\link{matrix.exponent}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
