% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\docType{package}
\name{GSVD-package}
\alias{GSVD}
\alias{GSVD-package}
\title{GSVD}
\description{
The generalized singular value decomposition (GSVD) generalizes the standard SVD (see \code{\link{svd}}) procedure through addition of (optional) constraints applied to the rows and/or columns of a matrix.
}
\details{
While dedicated to the GSVD, this package also includes some nice features that are helpful for matrix analyses. For example there are tests to check if matrices are empty (\code{\link{is.empty.matrix}}) or identity (\code{\link{is.identity.matrix}}); also included are operations such as generalized inverse (\code{\link{matrix.generalized.inverse}}) and matrix exponents (\code{\link{matrix.exponent}} or \code{\link{\%^\%}}).
}
\examples{
 ## an example of correspondence analysis.
 data(authors)
 Observed <- authors/sum(authors)
 row.w <- rowSums(Observed)
   row.W <- diag(1/row.w)
 col.w <- colSums(Observed)
   col.W <- diag(1/col.w)
 Expected <- row.w \%o\% col.w
 Deviations <- Observed - Expected
 ca.res <- gsvd(Deviations,row.W,col.W)


 # several examples of principal component analysis
 data(wine)
 wine.objective <- wine$objective
 ## "covariance" PCA
 cov.pca.data <- scale(wine.objective,scale=FALSE)
 cov.pca.res <- gsvd(cov.pca.data)
 ## "correlation" PCA
 cor.pca.data <- scale(wine.objective,scale=TRUE)
 cor.pca.res <- gsvd(cor.pca.data)
 ## "correlation" PCA with GSVD constraints
 cor.pca.res2 <- gsvd(cov.pca.data,RW=1/apply(wine.objective,2,var))


 # Three "two-table" technique examples
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)
 R <- t(X) \%*\% Y

 ## an example of partial least squares (correlation)
 pls.res <- gsvd(R)


 ## an example of canonical correlation analysis (CCA)
 ### NOTE:
 #### This is not "traditional" CCA because of the generalized inverse.
 #### However results are the same as standard CCA when data are not rank deficient.
 cca.res <- gsvd(
     DAT=(crossprod(X) \%^\% -1) \%*\% R \%*\% (crossprod(Y) \%^\% -1),
     LW=crossprod(X),
     RW=crossprod(Y)
 )
 cca.res$lx <- (X \%*\% cca.res$p)
 cca.res$ly <- (Y \%*\% cca.res$q)

 ## an example of reduced rank regression (RRR) a.k.a. redundancy analysis (RDA)
 ### NOTE:
 #### This is not "traditional" RRR because of the generalized inverse.
 #### However the results are the same as standard RRR when data are not rank deficient.
 rrr.res <- gsvd(
     DAT=(crossprod(X) \%^\% -1) \%*\% R,
     LW=crossprod(X)
 )
 rrr.res$lx <- (X \%*\% rrr.res$p)
 rrr.res$ly <- (Y \%*\% rrr.res$q)

}
\references{
Abdi, H. (2007). Singular Value Decomposition (SVD) and Generalized Singular Value Decomposition (GSVD). In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}.Thousand Oaks (CA): Sage. pp. 907-912.\cr
  Beaton, D., Fatt, C. R. C., & Abdi, H. (2014). An ExPosition of multivariate analysis with the singular value decomposition in R. \emph{Computational Statistics & Data Analysis}, 72, 176-189.
  Greenacre, M. (1984). Theory and applications of correspondence analysis. \emph{Academic Press.}\cr
  Holmes, S. (2008). Multivariate data analysis: the French way. \emph{In Probability and statistics: Essays in honor of David A. Freedman (pp. 219-233)}. Institute of Mathematical Statistics. \cr
  Holmes, S., & Josse, J. (2017). Discussion of “50 Years of Data Science”. \emph{Journal of Computational and Graphical Statistics}, 26(4), 768-769. \cr
  Husson, F., Josse, J., & Saporta, G. (2016). Jan de Leeuw and the French school of data analysis. \emph{Journal of Statistical Software}, 73(6), 1-17. \cr
  Lebart  L.,  Morineau  A., & Warwick  K.  (1984). Multivariate  Descriptive  Statistical  Analysis. \emph{J. Wiley}, New York. \cr
  Yanai, H., Takeuchi, K., & Takane, Y. (2011). Projection Matrices, Generalized Inverse Matrices, and Singular Value Decomposition. \emph{Springer-Verlag, New-York.}\cr
}
\seealso{
\code{\link{gsvd}}, \code{\link{tolerance.svd}}, \code{\link{\%^\%}}
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

}
\keyword{component}
\keyword{decomposition}
\keyword{genearlized}
\keyword{matrix}
\keyword{multivariate}
\keyword{orthogonal}
\keyword{svd}
\keyword{variance}
