% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invsqrt_psd_matrix}
\alias{invsqrt_psd_matrix}
\title{\code{invsqrt_psd_matrix}: inverse square root of a positive semi-definite matrix}
\usage{
invsqrt_psd_matrix(x)
}
\arguments{
\item{x}{A square matrix (presumably positive semi-definite)}
}
\value{
A matrix. The inverse square root of the \code{x} matrix
}
\description{
\code{invsqrt_psd_matrix} takes a square, positive semi-definite matrix and returns the inverse square root of that matrix (via the eigenvalue decomposition by way of \code{tolerance_eigen}).
}
\seealso{
\code{\link{tolerance_eigen}}
}
