% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.identical.matrix.R
\name{is.identical.matrix}
\alias{is.identical.matrix}
\title{\code{is.identical.matrix}: test if a matrix contains all identical values.}
\usage{
is.identical.matrix(x, tol = .Machine$double.eps, round.digits = 12)
}
\arguments{
\item{x}{A matrix to test.}

\item{tol}{Tolerance precision to eliminate all abs(x) values below \code{tol}. Default is \code{.Machine$double.eps}.}

\item{round.digits}{number of decimal places to round to (via \code{\link{round}} function).}
}
\value{
A boolean. TRUE if the matrix contains all identical values, FALSE if the matrix does not.
}
\description{
\code{is.identical.matrix} takes a matrix and tests if it contains all identical values.
}
