% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_diagonal_matrix}
\alias{is_diagonal_matrix}
\title{\code{is_diagonal_matrix}: test if a matrix is a diagonal matrix.}
\usage{
is_diagonal_matrix(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{A matrix to test.}

\item{tol}{Tolerance precision to eliminate all abs(x) values below \code{tol}. Default is \code{.Machine$double.eps}.}
}
\value{
A boolean. TRUE if the matrix is a diagonal matrix, FALSE if the matrix is not.
}
\description{
\code{is_diagonal_matrix} takes a matrix and tests if it is a diagonal matrix.
}
