% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods_helpers.R
\name{is_GSVD_gsvd}
\alias{is_GSVD_gsvd}
\title{is_GSVD_gsvd}
\usage{
is_GSVD_gsvd(x)
}
\arguments{
\item{x}{object to test}
}
\value{
boolean. \code{TRUE} if the object is of class gsvd, FALSE otherwise.
}
\description{
Tests if the \code{x} object is of class type "gsvd"
}
\details{
Only \code{\link{gsvd}} produces this class type.
}
\seealso{
\code{\link{inherits}}
}
