% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.exponent.R
\name{matrix.exponent}
\alias{matrix.exponent}
\title{\code{matrix.exponent}: raise matrix to a power and rebuild lower rank version}
\usage{
matrix.exponent(x, power = 1, k = 0, ...)
}
\arguments{
\item{x}{data matrix}

\item{power}{the power to raise \code{x} by (e.g., 2 is squared)}

\item{k}{the number of components to retain in order to build a lower rank estimate of \code{x}}

\item{...}{parameters to pass through to \code{\link{tolerance.svd}}}
}
\value{
The (possibly lower rank) raised to an arbitrary \code{power} version of \code{x}
}
\description{
\code{matrix.exponent} takes in a matrix and will compute raise that matrix to some arbitrary power via the singular value decomposition.
 Additionally, the matrix can be computed for a lower rank estimate of the matrix.
}
\examples{
 data(wine)
 X <- as.matrix(wine$objective)
 X.power_1 <- matrix.exponent(X)
 X / X.power_1

 ## other examples.
 X.power_2 <- matrix.exponent(X,power=2)
 X.power_negative.1.div.2 <- matrix.exponent(X,power=-1/2)

 X.power_negative.1 <- matrix.exponent(X,power=-1)
 X.power_negative.1 / (X \%^\% -1)

}
\seealso{
\code{\link{tolerance.svd}} and \code{\link{matrix.generalized.inverse}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
