% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.generalized.inverse.R
\name{matrix.generalized.inverse}
\alias{matrix.generalized.inverse}
\title{\code{matrix.generalized.inverse}: psuedo inverse and rebuild of a matrix}
\usage{
matrix.generalized.inverse(x, k = 0, ...)
}
\arguments{
\item{x}{data matrix to compute the pseudo-inverse of}

\item{k}{the number of components to retain in order to build a lower rank estimate of \code{x}}

\item{...}{parameters to pass through to \code{\link{tolerance.svd}}}
}
\value{
The (possibly lower rank) generalized inverse of \code{x}
}
\description{
\code{matrix.generalized.inverse} takes in a matrix and will compute the psuedo-inverse via the singular value decomposition.
 Additionally, the psuedo-inverse can be computed for a lower rank estimate of the matrix.
}
\examples{
 data(wine)
 X <- as.matrix(wine$objective)
 X.inv <- matrix.generalized.inverse(X)
 X.inv \%*\% X ## is approximately an identity.

}
\seealso{
\code{\link{tolerance.svd}} and \code{\link{matrix.exponent}}
}
\author{
Derek Beaton
}
\keyword{Moore-Penrose,}
\keyword{diagonalization,}
\keyword{eigen,}
\keyword{generalized}
\keyword{inverse}
\keyword{multivariate,}
\keyword{pseudo-inverse,}
