% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{synthetic_ADNI}
\alias{synthetic_ADNI}
\title{synthetic_ADNI}
\format{
A matrix that contains 623 observations (rows) and 17 variables (columns) of various data types (i.e., a mixture of continuous, categorical, and ordinal)
}
\usage{
synthetic_ADNI
}
\description{
A synthetic data set derived from the Alzheimer's Disease Neuroimaging Initiative (ADNI). Data were generated from a specific set of variables in the \code{ADNIMERGE} \code{R} package. Synthetic data were produced with the \code{synthpop} package with the "cart" method.
}
\details{
see http://adni.loni.usc.edu/ for more details on the ADNI data and the \code{ADNIMERGE} package
}
\keyword{datasets}
