% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{synthetic_ONDRI}
\alias{synthetic_ONDRI}
\title{synthetic_ONDRI}
\format{
A matrix that contains 138 observations (rows) and 17 variables (columns) of various data types (i.e., a mixture of continuous, categorical, and ordinal), including some identification variables (e.g., IDs)
}
\usage{
synthetic_ONDRI
}
\description{
A synthetic data set derived from the Ontario Neurodegenerative Disease Research Initiative (ONDRI). Data were generated from a specific set of variables in the \code{ADNIMERGE} \code{R} package. Synthetic data were produced with the \code{synthpop} package with the "cart" method.
}
\details{
see http://ondri.ca for more details on the study, and see https://github.com/ondri-nibs/toy_data for details on the synthetic data. To note, the \code{data.frame} available here is in a more "prepared" format than available as raw data files.
}
\keyword{datasets}
