% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.eigen.R
\name{tolerance.eigen}
\alias{tolerance.eigen}
\title{\code{tolerance.eigen}: An SVD to truncate potentially spurious (near machine precision) components.}
\usage{
tolerance.eigen(x, tol = .Machine$double.eps, ...)
}
\arguments{
\item{x}{A data matrix of size for input to the singular value decomposition (\code{\link{svd}})}

\item{tol}{Default is \code{.Machine$double.eps}. A tolerance level for eliminating near machine precision components.
Use of this parameter causes \code{tolerance.eigen} to stop if negative or complex singular values are detected.
The use of \code{tol < 0}, \code{NA}, \code{NaN}, \code{Inf}, \code{-Inf}, or \code{NULL} passes through to \code{\link{svd}}.}

\item{nu}{The number of left singular vectors to be computed. Default is \code{min(dim(x))}}

\item{nv}{The number of right singular vectors to be computed. Default is \code{min(dim(x))}}
}
\value{
A list with three elements (like \code{svd}):
 \item{values}{ A vector containing the eigen values of x > \code{tol}.}
 \item{vectors}{ A matrix whose columns contain the right singular vectors of x, present if nv > 0. Dimension \code{min(c(ncol(x), nv, length(d))}.}
}
\description{
\code{tolerance.eigen} eliminates likely spurious components: any eigenvalue (squared singular value) below a tolerance level is elminated.
   The (likely) spurious singular values and vectors are then eliminated from \code{$u}, \code{$d}, and \code{$v}.
   Additionally, all values in \code{abs($u)} or \code{abs($v)} that fall below the \code{tol} are set to 0.
   The use of a real positive value for \code{tol} will eliminate any small valued components.
   With \code{tol}, \code{tolerance.eigen} will stop if any singular values are complex or negative.
}
\seealso{
\code{\link{eigen}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
