% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.svd.R
\name{tolerance.svd}
\alias{tolerance.svd}
\title{\code{tolerance.svd}: automatically truncate spurious (tiny variance or negative or imaginary) components.}
\usage{
tolerance.svd(x, nu = min(dim(x)), nv = min(dim(x)),
  tol = .Machine$double.eps)
}
\arguments{
\item{x}{A data matrix of size for input to the singular value decomposition (\code{\link{svd}})}

\item{nu}{The number of left singular vectors to be computed. Default is \code{min(dim(x))}}

\item{nv}{The number of right singular vectors to be computed. Default is \code{min(dim(x))}}

\item{tol}{Default is \code{.Machine$double.eps}. A parameter with two roles: A tolerance level for 1: eliminating (tiny variance or negative or imaginary) components and 2: converting all values < tol to 0 in \code{u} and \code{v}}
}
\value{
A list with three elements (like \code{svd}):
 \item{d}{ A vector containing the singular values of x > \code{tol}.}
 \item{u}{ A matrix whose columns contain the left singular vectors of x, present if nu > 0. Dimension \code{min(c(nrow(x), nu, length(d))}.}
 \item{v}{ A matrix whose columns contain the right singular vectors of x, present if nv > 0. Dimension \code{min(c(ncol(x), nv, length(d))}.}
}
\description{
\code{tolerance.svd} eliminates likely spurious components: any eigenvalue (squared singular value) below a tolerance level is elminated.
   The (likely) spurious singular values and vectors are then eliminated from \code{$u}, \code{$d}, and \code{$v}.
   Additionally, all values in \code{abs($u)} or \code{abs($v)} that fall below the \code{tol} are set to 0.
}
\examples{
 data(wine)
 X <- scale(as.matrix(wine$objective))
 s_asis <- tolerance.svd(X)
 s_sqrt.Machine <- tolerance.svd(X,tol=sqrt(.Machine$double.eps))
 s_000001 <- tolerance.svd(X,tol=.000001)

}
\seealso{
\code{\link{svd}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
