% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{wine}
\alias{wine}
\title{Objective and subjective wine data}
\format{
A list called 'wine' with three matrices:
\describe{
\item{objective}{a 36 wine x 5 attribute matrix of objective variables for wines.}
\item{subjective}{a 36 wine x 9 attribute matrix of subjective variables for wines.}
\item{supplemental}{a 36 wine x 3 attribute matrix of supplemental information on the wines (country, color, grape).}
}
}
\usage{
wine
}
\description{
Data used for illustrative purposes in Abdi H., Guillemot, V., Eslami, A., & Beaton, D. (in Press, 2018). Canonical correlation analysis (CCA). In R. Alhajj and J. Rokne (Eds.), Encyclopedia of Social Networks and Mining (2nd Edition). New York: Springer Verlag.
}
\keyword{datasets}
