% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geigen.R
\name{geigen}
\alias{geigen}
\title{Generalized eigenvalue decomposition}
\usage{
geigen(DAT, W, k = 0, tol = .Machine$double.eps, symmetric)
}
\arguments{
\item{DAT}{a square, symmetric data matrix to decompose}

\item{W}{\bold{W}eights -- the constraints applied to the matrix and thus the eigen vectors.}

\item{k}{total number of components to return though the full variance will still be returned (see \code{d.orig}). If 0, the full set of components are returned.}

\item{tol}{default is .Machine$double.eps. A parameter with two roles: A tolerance level for (1) eliminating (tiny variance or negative or imaginary) components and (2) converting all values < tol to 0 in \code{v}.}

\item{symmetric}{if \code{DAT} is symmetric, set as TRUE. See \code{\link{eigen}}.}
}
\value{
A list with eight elements:
\item{d.orig}{A vector containing the singular values of DAT above the tolerance threshold (based on eigenvalues).}
\item{l.orig}{A vector containing the eigen values of DAT above the tolerance threshold (\code{tol}).}
\item{tau}{A vector that contains the (original) explained variance per component (via eigenvalues: \code{$l.orig}.}
\item{d}{A vector of length \code{min(length(d.orig), k)} containing the retained singular values of DAT}
\item{l}{A vector of length \code{min(length(l.orig), k)} containing the retained eigen values of DAT}
\item{v}{Eigenvectors. Dimensions are \code{ncol(DAT)} by k.}
\item{q}{Generalized eigenvectors. Dimensions are \code{ncol(DAT)} by k.}
\item{fj}{Component scores. Dimensions are \code{ncol(DAT)} by k.}
}
\description{
\code{geigen} takes in constraints (\code{W}), (usually diagonal matrices, but any positive semi-definite matrix) that are applied to the data (\code{DAT}).
  Constraints are used for the orthogonality conditions.
}
\examples{
Observed <- authors/sum(authors)
row.w <- rowSums(Observed)
row.W <- diag(1/row.w)
col.w <- colSums(Observed)
col.W <- diag(1/col.w)
Expected <- row.w \%o\% col.w
Deviations <- Observed - Expected
ca.res_gsvd <- gsvd(Deviations,row.W,col.W)
ca.res_geigen <- geigen(t(Deviations) \%*\% diag(1/row.w) \%*\% Deviations, col.W)

}
\seealso{
\code{\link{tolerance.eigen}}, \code{\link{tolerance.svd}}, \code{\link{gsvd}} and \code{\link{svd}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
