% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\name{gsvd}
\alias{gsvd}
\title{Generalized SVD}
\usage{
gsvd(DAT, LW, RW, k = 0, tol = .Machine$double.eps)
}
\arguments{
\item{DAT}{a data matrix to decompose}

\item{LW}{\bold{L}eft \bold{W}eights -- the constraints applied to the left side (rows) of the matrix and thus left singular vectors.}

\item{RW}{\bold{R}ight \bold{W}eights -- the constraints applied to the right side (rows) of the matrix and thus right singular vectors.}

\item{k}{total number of components to return though the full variance will still be returned (see \code{d.orig}). If 0, the full set of components are returned.}

\item{tol}{default is .Machine$double.eps. A parameter with two roles: A tolerance level for (1) eliminating (tiny variance or negative or imaginary) components and (2) converting all values < tol to 0 in \code{u} and \code{v}.}
}
\value{
A list with nine elements:
\item{d.orig}{A vector containing the singular values of DAT > \code{tol}.}
\item{tau}{A vector that contains the (original) explained variance per component (eigenvalues derived from \code{$d.orig}.}
\item{d}{A vector containing the singular values of x > \code{tol}. Length is \code{min(length(d.orig), k)}}
\item{u}{Left (rows) singular vectors. Dimensions are \code{nrow(DAT)} by k.}
\item{p}{Left (rows) generalized singular vectors. Dimensions are \code{nrow(DAT)} by k.}
\item{fi}{Left (rows) component scores. Dimensions are \code{nrow(DAT)} by k.}
\item{v}{Right (columns) singular vectors. Dimensions are \code{ncol(DAT)} by k.}
\item{q}{Right (columns) generalized singular vectors. Dimensions are \code{ncol(DAT)} by k.}
\item{fj}{Right (columns) component scores. Dimensions are \code{ncol(DAT)} by k.}
}
\description{
\code{gsvd} takes in left (\code{LW}) and right (\code{RW}) constraints (usually diagonal matrices, but any positive semi-definite matrix is fine) that are applied to the data (\code{DAT})
  Left and right constraints are used for the orthogonality conditions.
}
\examples{

 ## an example of correspondence analysis.
 data(authors)
 Observed <- authors/sum(authors)
 row.w <- rowSums(Observed)
   row.W <- diag(1/row.w)
 col.w <- colSums(Observed)
   col.W <- diag(1/col.w)
 Expected <- row.w \%o\% col.w
 Deviations <- Observed - Expected
 ca.res <- gsvd(Deviations,row.W,col.W)


 # several examples of principal component analysis
 data(wine)
 wine.objective <- wine$objective
 ## "covariance" PCA
 cov.pca.data <- scale(wine.objective,scale=FALSE)
 cov.pca.res <- gsvd(cov.pca.data)
 ## "correlation" PCA
 cor.pca.data <- scale(wine.objective,scale=TRUE)
 cor.pca.res <- gsvd(cor.pca.data)
 ## "correlation" PCA with GSVD constraints
 cor.pca.res2 <- gsvd(cov.pca.data,RW=1/apply(wine.objective,2,var))


 # Three "two-table" technique examples
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)
 R <- t(X) \%*\% Y

 ## an example of partial least squares (correlation)
 pls.res <- gsvd(R)


 ## an example of canonical correlation analysis (CCA)
 ### NOTE:
 #### This is not "traditional" CCA because of the generalized inverse.
 #### However the results are the same as standard CCA when data are not rank deficient.
 cca.res <- gsvd(
     DAT=(crossprod(X) \%^\% -1) \%*\% R \%*\% (crossprod(Y) \%^\% -1),
     LW=crossprod(X),
     RW=crossprod(Y)
 )
 cca.res$lx <- (X \%*\% cca.res$p)
 cca.res$ly <- (Y \%*\% cca.res$q)


 ## an example of reduced rank regression (RRR) a.k.a. redundancy analysis (RDA)
 ### NOTE:
 #### This is not "traditional" RRR because of the generalized inverse.
 #### However the results are the same as standard RRR when data are not rank deficient.
 rrr.res <- gsvd(
     DAT=(crossprod(X) \%^\% -1) \%*\% R,
     LW=crossprod(X)
 )
 rrr.res$lx <- (X \%*\% rrr.res$p)
 rrr.res$ly <- (Y \%*\% rrr.res$q)

}
\seealso{
\code{\link{tolerance.svd}} and \code{\link{svd}}
}
\author{
Derek Beaton
}
\keyword{diagonalization,}
\keyword{eigen}
\keyword{multivariate,}
